Created restock room and updated restock feature
To restock a shelf, players must go to the restock room and collect an item from the shelf with the respective color. After a restock, the player must grab another item from the restock room in order to restock again.

Restock amount can be adjusted to make some shelves harder to restock, possibly for shelves that earn more money or with fewer max stock.